Use u64 instead of u32 to not limit scans to below 42GHz

+++ src/lib.rs
@@ -15,8 +15,8 @@ use colors::Palettes;
 struct Measurement {
     date: String,
     time: String,
-    freq_low: u32,
-    freq_high: u32,
+    freq_low: u64,
+    freq_high: u64,
     freq_step: f64,
     samples: u32,
     values: Vec<f32>,
